chore: no more v2 for operator related services
